refactor: add CoroutineDispatcherModule
