+++ app/src/main/kotlin/com/b4tchkn/times/di/CoroutineDispatcherModule.kt
@@ -1,0 +1,21 @@
+package com.b4tchkn.times.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Qualifier
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+@Module
+@InstallIn(SingletonComponent::class)
+object CoroutineDispatcherModule {
+    @IODispatcher
+    @Provides
+    fun provideCoroutineIODispatcher(): CoroutineDispatcher = Dispatchers.IO
+}
+
+@Retention(AnnotationRetention.BINARY)
+@Qualifier
+annotation class IODispatcher
